requirements.txt
langchain
openai
chromadb
pypdf
sentence-transformers
fastapi
uvicorn
python-dotenv
scikit-learn
